feat: Suspense-based streaming SSR for signal feed route
Streams the page shell immediately while signal data loads server-side,
showing the existing skeleton during the streamed-in period.

- app/app/loading.tsx: instant skeleton shell for the route segment
- components/signal/SignalFeedServer.tsx: async server component that
  fetches page 1 via buildSignalPage and passes it as initialData
- app/app/page.tsx: converted to server component; wraps SignalFeedServer
  in Suspense with skeleton fallback
- components/AppShell.tsx: extracted all client logic (wallet, modals,
  trade card) so page.tsx can be a server component
- components/signal/SignalFeed.tsx: accepts initialData prop, seeds
  useInfiniteQuery cache to eliminate the client waterfall

Closes #257
