+++ components/signal/SignalFeed.tsx
@@ -31,7 +31,12 @@ const PAGE_SIZE = 10;
 // #98: 5-minute stale time so recently-viewed pages are served from cache
 const STALE_TIME = 1000 * 60 * 5;
 
-export function SignalFeed() {
+interface SignalFeedProps {
+  /** Server-fetched first page — eliminates the client waterfall on initial load */
+  initialData?: SignalResponse;
+}
+
+export function SignalFeed({ initialData }: SignalFeedProps = {}) {
   const feedRef = useRef<HTMLDivElement | null>(null);
 
   // #99: provider search state (persisted in filter store)
@@ -73,6 +78,13 @@ export function SignalFeed() {
     initialPageParam: 1,
     staleTime: STALE_TIME,
     placeholderData: (prev) => prev,
+    // Seed the cache with the server-fetched first page so no client waterfall occurs
+    ...(initialData && {
+      initialData: {
+        pages: [initialData],
+        pageParams: [1],
+      },
+    }),
   });
 
   const allSignals = useMemo<Signal[]>(
